feat: add inventory, split gameplay scene, and refactor core systems
- Add inventory.c to systems module
- Split gameplay scene into update.c and draw.c
- Add IVY_SaveManager_Flush for Windows (FlushViewOfFile/FlushFileBuffers)
- Refactor IvyKeybind API to accept SaveManager for load/update/reset
- Add default keybind mapping in SaveHeader init
- Add IVY_SLOT_EMPTY and IVY_FONT_SIZE constants to types.h
- Fix IVY_CHECK macro signature (variadic args)
- Add IVY_CATEGORY_MAX and rename IVY_TYPE_* enums to IVY_ITEM_TYPE_*
- Add -Wl,--start-group/--end-group linker flags to fix circular deps
- Reduce IvyGame arena size comment from 64 to 32

+++ src/graphics/collusion.c
@@ -24,7 +24,7 @@ typedef struct {
 static IvyCollisionVertex *CollisionMap_LoadVertices(IvyAssetManager *restrict manager, IvyArenaLinear *restrict arena, int *outVertexCount)
 {
     usize dataSize;
-    const float *rawData = (const float *)Ivy_Asset_Get(manager, ASSET_MAPS_MAP_1_COLLISION_BIN, &dataSize);
+    const u8 *rawData = (const u8 *)Ivy_Asset_Get(manager, ASSET_MAPS_MAP_1_COLLISION_BIN, &dataSize);
 
     IVY_CHECK(rawData != NULL, "[CollisionMap] Failed to load collusion binary");
